LangChain/chatbot/requirements.txt
gradio
langchain
langchain-chroma
langchain-community
langchain-openai
python-dotenv
chromadb
pypdf
requests
